fix(diff): compute git diff in a better way (#41)
This change uses the output of `git diff` to calculate the diff and it
checks the lines for +/- to count added and removed lines in the status.
It also fixes the code to diff against the base commit of the branch.
Previously, there was an issue where we would not diff against the base
properly and pausing/resuming would reset the stats. Now, the diff stays
even if the changes are committed.