fix: `c` Command Checkout Locally Only (#123)
**Problem**
Contrary to what's highlighted in the README
> Review changes before applying them, checkout changes before pushing
them

You cannot checkout changes locally before pushing them. The `c` command
to checkout pushes your changes to GitHub first before allowing you to
checkout locally.

Push functionality already exists with the `p` command, so it makes
sense to have `c` only commit to a branch locally and not push.

This pull request addresses this bug with the `c` command to only commit
locally and not push to GitHub.

Resolves 
- https://github.com/smtg-ai/claude-squad/issues/122
- https://github.com/smtg-ai/claude-squad/issues/68